refactor: team mutations type 분리
